requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
# Embeddings + Vector DB
sentence-transformers>=2.6.0
faiss-cpu==1.7.4
tqdm==4.66.1
opencv-python==4.5.5.64
numpy == 1.23.5
tensorflow-cpu == 2.12.0
psycopg2-binary==2.9.7
# API Dependencies
fastapi==0.68.0
uvicorn
python-multipart